ssss

+++ ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: rufaccia <rufaccia@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/22 18:23:02 by rufaccia          #+#    #+#             */
-/*   Updated: 2024/05/14 12:41:41 by rufaccia         ###   ########.fr       */
+/*   Updated: 2024/05/15 09:51:47 by rufaccia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	size_t	i;
